create psascan.cc for psascan.h because it's causing strange behaviors for llvm backend and it slows down the compiling progress a lot

+++ psascan/sa_use.cc
@@ -3,6 +3,7 @@
 
 #include <iostream>
 #include <string>
+#include <omp.h>
 
 void sa_build(const char *fname, long ram_use) {
 	std::string t_fname(fname);
@@ -11,8 +12,7 @@ void sa_build(const char *fname, long ram_use) {
 
 //	long ram_use = 3072L << 20;
 //	long ram_use = 8L << 30;
-	long max_threads = (long)omp_get_max_threads();
-
+	long max_threads = omp_get_max_threads();
 	pSAscan(t_fname, out_fname, gap_fname,
 	        ram_use, max_threads, false);
 }
